last_change

+++ eTickets/Controllers/CinemasController.cs
@@ -1,102 +1,92 @@
-﻿using eTickets.Data.Services;
+﻿using eTickets.Data.Enums;
+using eTickets.Data.Services;
 using eTickets.Models;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Formatters;
+using Microsoft.EntityFrameworkCore.Metadata.Internal;
 
 namespace eTickets.Controllers
 {
     public class CinemasController : Controller
     {
+        //inject the data 1
+        //when we create the service we injected there so we remove injection from here
         private readonly ICinemasService _service;
 
         public CinemasController(ICinemasService service)
         {
             _service = service;
         }
-
-        // GET: Cinemas
         public async Task<IActionResult> Index()
         {
-            var cinemas = await _service.GetAllAsync();
-            return View(cinemas);
+            //also change here cause we use service now not _contex from dbcontext
+            var data = await _service.GetAllAsync();
+            return View(data);
         }
-
-        // GET: Cinemas/Details/5
-        public async Task<IActionResult> Details(int id)
-        {
-            var cinema = await _service.GetByIdAsync(id);
-            if (cinema == null)
-            {
-                return View("NotFound");
-            }
-            return View(cinema);
-        }
-
-        // GET: Cinemas/Create
+        //Get: Actors/create
         public IActionResult Create()
         {
             return View();
         }
-
-        // POST: Cinemas/Create
         [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Name,Location")] Cinema cinema)
+        public async Task<IActionResult> Create([Bind("Name, Logo, Description")] Cinema cinema)
         {
-            if (ModelState.IsValid)
-            {
+            
+            
                 await _service.AddAsync(cinema);
                 return RedirectToAction(nameof(Index));
+            
+            
+        }
+        public async Task<IActionResult> Details(int id)
+        {
+            var cinemaDetails = await _service.GetByIdAsync(id);
+            if (cinemaDetails == null)
+            {
+                return View("NotFound");
             }
-            return View(cinema);
+
+            return View(cinemaDetails);
         }
 
-        // GET: Cinemas/Edit/5
+        //Get: Actors/Edit
         public async Task<IActionResult> Edit(int id)
         {
-            var cinema = await _service.GetByIdAsync(id);
-            if (cinema == null)
+            var actorDetails = await _service.GetByIdAsync(id);
+            if (actorDetails == null)
             {
                 return View("NotFound");
             }
-            return View(cinema);
+            return View(actorDetails);
         }
-
-        // POST: Cinemas/Edit/5
         [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Name,Location")] Cinema cinema)
+        public async Task<IActionResult> Edit(int id, [Bind("Id, Name, Logo, Description")] Cinema cinema)
         {
-            if (id != cinema.Id)
-            {
-                return View("NotFound");
-            }
-
-            if (ModelState.IsValid)
-            {
                 await _service.UpdateAsync(id, cinema);
                 return RedirectToAction(nameof(Index));
-            }
-            return View(cinema);
         }
 
-        // GET: Cinemas/Delete/5
+        //Get: Actors/Delete
         public async Task<IActionResult> Delete(int id)
         {
-            var cinema = await _service.GetByIdAsync(id);
-            if (cinema == null)
+            var cinemaDetails = await _service.GetByIdAsync(id);
+            if (cinemaDetails == null)
             {
                 return View("NotFound");
             }
-            return View(cinema);
+            return View(cinemaDetails);
         }
-
-        // POST: Cinemas/Delete/5
         [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            var cinemaDetails = await _service.GetByIdAsync(id);
+            if (cinemaDetails == null)
+            {
+                return View("NotFound");
+            }
             await _service.DeleteAsync(id);
             return RedirectToAction(nameof(Index));
         }
+
     }
 }
